src/requirements.txt
python-dotenv
aiohttp
teams-ai>=1.6.0,<2.0.0
azure-ai-projects>=1.0.0b11
azure-identity>=1.15.0
azure-core>=1.29.0